splitted js
